feat: 🎸 app share
