+++ src/components/Loading/index.tsx
@@ -1,27 +1,29 @@
 import { useLottie } from 'lottie-react';
 import LoadingAnimation from 'assets/img/loading-animation.json';
 import LoadingAnimationBlue from 'assets/img/loading-animation-blue.json';
+import { useMemo } from 'react';
+import React from 'react';
+
+const loadingImg = {
+  white: LoadingAnimation,
+  blue: LoadingAnimationBlue,
+};
 
 interface IProps {
   color?: 'white' | 'blue';
 }
 
-export default function Loading({ color = 'blue' }: IProps) {
-  const loadingImg = {
-    white: LoadingAnimation,
-    blue: LoadingAnimationBlue,
-  };
-  const Animation = () => {
-    const options = {
+function Loading({ color = 'blue' }: IProps) {
+  const options = useMemo(() => {
+    return {
       animationData: loadingImg[color],
       loop: true,
       autoplay: true,
     };
+  }, [color]);
 
-    const { View } = useLottie(options, { width: '40px', height: '40px' });
-
-    return View;
-  };
-
-  return <Animation />;
+  const { View } = useLottie(options, { width: '40px', height: '40px' });
+  return View;
 }
+
+export default React.memo(Loading);
